Deliveries Report added. Calculus bug in Building, Commercial and Planning fixed.

+++ libraries/temporary_area/building/tmp_ar_building.py
@@ -146,7 +146,7 @@ def tmp_ar_building(stg_consolidado_corte, tbl_proyectos):
 
     #Proceed with calculations of the equation above
 
-    auxCol['delta']=(auxCol['stg_duracion_cantidad']*(1-(auxCol['tpc_avance_cc']/100)))
+    auxCol['delta']=(auxCol['stg_duracion_cantidad']*(1-(auxCol['tpc_avance_cc']/100))).astype(int)
     auxCol['delta_days'] = auxCol['delta'].apply(np.ceil).apply(lambda x: pd.Timedelta(x, unit='D'))
     auxCol['tpc_fin_proyectado_pesimista']=auxCol['tpc_fin_proyectado_optimista']+auxCol['delta_days']
